Add stubs for authentication commands
