+++ internal/handler/registry/hana.go
@@ -44,6 +44,7 @@ func init() {
 
 			DisableFilterPushdown: opts.DisableFilterPushdown,
 			EnableOplog:           opts.EnableOplog,
+			EnableNewAuth:         opts.EnableNewAuth,
 		}
 
 		h, err := handler.New(handlerOpts)
